requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
cachetools==5.5.1
ccxt==4.4.61
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
convertdate==2.4.0
cryptography==44.0.0
dateparser==1.2.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.8
frozenlist==1.5.0
google-auth==2.38.0
google-genai==1.1.0
h11==0.14.0
holidays==0.66
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lunardate==0.2.2
lxml==5.3.0
multidict==6.1.0
numpy==1.26.4
openai==1.61.1
outcome==1.3.0.post0
pandas==2.2.3
pandas-ta==0.3.14b0
pillow==11.1.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyluach==2.2.0
PyMeeus==0.5.12
PySocks==1.7.1
python-binance==1.0.27
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2025.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.45.3
ta==0.11.0
ta-lib==0.6.3
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
websockets==14.2
workalendar==17.0.0
wsproto==1.2.0
yarl==1.18.3
